requirements.txt
streamlit
requests
mysql-connector-python
python-dotenv
matplotlib
numpy
pandas
plotly
